Fix / improve types

+++ freqtrade/freqtradebot.py
@@ -141,7 +141,7 @@ class FreqtradeBot(LoggingMixin):
         # Initialize protections AFTER bot start - otherwise parameters are not loaded.
         self.protections = ProtectionManager(self.config, self.strategy.protections)
 
-        def log_took_too_long(duration: int, time_limit: int):
+        def log_took_too_long(duration: float, time_limit: float):
             logger.warning(
                 f"Strategy analysis took {duration:.2f}, which is 25% of the timeframe. "
                 "This can lead to delayed orders and missed signals."
